# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/google": {
      "post": {
        "operationId": "AuthController_googleAuth",
        "summary": "Authenticate with Google ID token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid Google token"
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "summary": "Get current authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/challenges": {
      "post": {
        "operationId": "ChallengesController_create",
        "summary": "Create a new challenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Challenge created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/{id}": {
      "delete": {
        "operationId": "ChallengesController_delete",
        "summary": "Delete a challenge (creator only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only the creator can delete"
          },
          "404": {
            "description": "Challenge not found"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ChallengesController_update",
        "summary": "Update a challenge (creator only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChallengeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Only the creator can update"
          },
          "404": {
            "description": "Challenge not found"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ChallengesController_findOne",
        "summary": "Get a challenge by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Challenge details"
          },
          "404": {
            "description": "Challenge not found"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/{id}/checkin": {
      "patch": {
        "operationId": "ChallengesController_updateCheckin",
        "summary": "Update today's check-in for a challenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCheckinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Check-in updated"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Must join challenge first"
          },
          "404": {
            "description": "Challenge not found or no check-in for today"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ChallengesController_checkin",
        "summary": "Check in for a challenge (daily progress)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Check-in recorded"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Must join challenge first"
          },
          "404": {
            "description": "Challenge not found"
          },
          "409": {
            "description": "Already checked in for this date"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/{id}/join": {
      "post": {
        "operationId": "ChallengesController_join",
        "summary": "Join a challenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully joined the challenge"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Challenge not found"
          },
          "409": {
            "description": "Already a member"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/explore": {
      "get": {
        "operationId": "ChallengesController_explore",
        "summary": "Browse public challenges sorted by popularity",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of public challenges"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/my": {
      "get": {
        "operationId": "ChallengesController_getMyChallenges",
        "summary": "Get all challenges the current user is a member of",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of user challenges"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/{id}/checkins": {
      "get": {
        "operationId": "ChallengesController_getCheckins",
        "summary": "Get challenge check-ins (paginated)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of check-ins"
          },
          "403": {
            "description": "No access to this challenge"
          },
          "404": {
            "description": "Challenge not found"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/challenges/{id}/members": {
      "get": {
        "operationId": "ChallengesController_getMembers",
        "summary": "Get challenge members with streaks (paginated)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "best_streak",
              "enum": [
                "best_streak",
                "current_streak",
                "joined_at"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of members"
          },
          "404": {
            "description": "Challenge not found"
          }
        },
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/dev/reset-user-progress": {
      "post": {
        "operationId": "DevController_resetUserProgress",
        "summary": "[DEV ONLY] Reset progress for test user (teammates, check-ins)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User progress reset"
          },
          "403": {
            "description": "Only available in development"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Dev"
        ]
      }
    }
  },
  "info": {
    "title": "Challenges.me API",
    "description": "Social habit and challenge platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GoogleAuthDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "Google ID token obtained from the mobile client",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6..."
          }
        },
        "required": [
          "idToken"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "skip_build_team": {
            "type": "boolean",
            "description": "When true, skip \"build your team\" flow and go directly to join pack"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "picture",
          "skip_build_team"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
...
      "CreateChallengeDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Challenge title",
            "example": "30 Days of Meditation"
          },
          "description": {
            "type": "string",
            "description": "Challenge description"
          },
          "imageUrl": {
            "type": "string",
            "description": "Challenge image URL"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "teammates_only",
              "private_invite"
            ],
            "default": "public"
          },
          "type": {
            "type": "string",
            "enum": [
              "strict",
              "flexible"
            ],
            "example": "strict"
          },
          "media_requirement": {
            "type": "string",
            "enum": [
              "none",
              "image_or_video_required",
              "image_required",
              "video_required"
            ],
            "default": "none"
          },
          "start_date": {
            "type": "string",
            "description": "Start date (YYYY-MM-DD)",
            "example": "2026-03-01"
          },
          "end_date": {
            "type": "string",
            "description": "End date (YYYY-MM-DD)",
            "example": "2026-03-31"
          }
        },
        "required": [
          "title",
          "visibility",
          "type",
          "media_requirement",
          "start_date"
        ]
      },
      "UpdateChallengeDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Challenge title"
          },
          "description": {
            "type": "string",
            "description": "Challenge description"
          },
          "imageUrl": {
            "type": "string",
            "description": "Challenge image URL"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "teammates_only",
              "private_invite"
            ]
          },
          "media_requirement": {
            "type": "string",
            "enum": [
              "none",
              "image_or_video_required",
              "image_required",
              "video_required"
            ]
          },
          "end_date": {
            "type": "string",
            "description": "End date (YYYY-MM-DD)"
          }
        }
      },
      "CheckinMediaItemDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ]
          },
          "url": {
            "type": "string",
            "description": "URL of the uploaded image or video"
          }
        },
        "required": [
          "type",
          "url"
        ]
      },
      "UpdateCheckinDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "failed"
            ],
            "description": "success or failed"
          },
          "text": {
            "type": "string",
            "description": "Optional text note"
          },
          "media": {
            "description": "Replace media. Each item: { type: \"image\" | \"video\", url: string }",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckinMediaItemDto"
            }
          }
        }
      },
      "CreateCheckinDto": {
        "type": "object",
        "properties": {
          "checkin_date": {
            "type": "string",
            "description": "Date of the check-in (YYYY-MM-DD). Defaults to today.",
            "example": "2026-03-15"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "failed"
            ],
            "description": "success or failed"
          },
          "text": {
            "type": "string",
            "description": "Optional text note"
          },
          "media": {
            "description": "Optional media (images/videos) attached to the check-in",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckinMediaItemDto"
            }
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}